refactor(tui): move task-panel sticky strip render into agent_tui_kit (Ctx slice)
