+++ crates/agent_tui_kit/src/render/mod.rs
@@ -17,4 +17,5 @@ pub mod render_md;
 pub mod renderable;
 pub mod selectable_text;
 pub mod slash_style;
+pub mod task_panel;
 pub mod util;
