+++ crates/tui/src/render/task_panel.rs
@@ -1,20 +1,10 @@
-//! Persistent tasks sticky strip under Log.
+//! Persistent tasks sticky strip — app-layer wrapper (mouse area + kit render).
 
-use ratatui::{
-    Frame,
-    layout::Rect,
-    style::{Modifier, Style},
-    text::{Line, Span},
-    widgets::{Block, Borders, Clear, Paragraph},
-};
+use ratatui::{Frame, layout::Rect};
 
-use crate::widgets::state::{
-    App,
-    task_panel::{format_grouped_lines, format_sticky_title_line},
-};
+use crate::widgets::state::App;
 
-/// Extra rows for sticky chrome: bottom border joins the Log box (sides continue).
-pub(crate) const STICKY_BORDER_ROWS: u16 = 1;
+pub(crate) use agent_tui_kit::render::task_panel::STICKY_BORDER_ROWS;
 
 pub(crate) fn sticky_host_visible(app: &App) -> bool {
     app.task_panel.visible
@@ -29,7 +19,7 @@ pub(crate) fn sticky_host_content_height(app: &App) -> usize {
         return 1;
     }
     // title + hairline + body
-    let body = format_grouped_lines(
+    let body = crate::widgets::state::task_panel::format_grouped_lines(
         &app.task_panel.snapshot,
         app.task_panel.scroll,
         app.task_panel.max_visible,
@@ -41,99 +31,8 @@ pub(crate) fn sticky_host_content_height(app: &App) -> usize {
 
 pub(crate) fn render_task_panel(frame: &mut Frame, area: Rect, app: &mut App) {
     app.mouse.task_panel_area = area;
-    if area.height == 0 || area.width == 0 || !sticky_host_visible(app) {
-        return;
-    }
-
-    frame.render_widget(Clear, area);
-
-    let block = Block::default()
-        .borders(Borders::LEFT | Borders::RIGHT | Borders::BOTTOM)
-        .border_type(app.theme.block_border_type())
-        .border_style(Style::default().fg(app.theme.border))
-        .style(Style::default().bg(app.theme.bg));
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-    frame.render_widget(
-        Paragraph::new("").style(Style::default().bg(app.theme.bg)),
-        inner,
-    );
-
-    if inner.height == 0 || inner.width == 0 {
-        return;
-    }
-
-    let bg = app.theme.bg;
-    let accent = Style::default()
-        .fg(app.theme.accent)
-        .bg(bg)
-        .add_modifier(Modifier::BOLD);
-    let muted = Style::default().fg(app.theme.muted_fg()).bg(bg);
-    let row_style = Style::default().fg(app.theme.fg).bg(bg);
-
-    let msgs = app.msgs();
-    let rest = format_sticky_title_line(&msgs, &app.task_panel.snapshot)
-        .trim_start_matches('▸')
-        .trim_start()
-        .to_string();
-    let title_spans = vec![
-        Span::styled("[Tasks]", accent),
-        Span::styled(" ", row_style),
-        Span::styled(rest, row_style),
-    ];
-    let title_line = Line::from(title_spans);
-
-    if !app.task_panel.expanded || inner.height == 1 {
-        frame.render_widget(Paragraph::new(title_line), inner);
-        return;
-    }
-
-    // Expanded: title + hairline + body
-    let title_area = Rect {
-        x: inner.x,
-        y: inner.y,
-        width: inner.width,
-        height: 1,
-    };
-    frame.render_widget(Paragraph::new(title_line), title_area);
-
-    if inner.height < 2 {
-        return;
-    }
-    let gap_area = Rect {
-        x: inner.x,
-        y: inner.y.saturating_add(1),
-        width: inner.width,
-        height: 1,
-    };
-    let rule = "─".repeat(inner.width as usize);
-    frame.render_widget(
-        Paragraph::new(Line::from(Span::styled(rule, muted))),
-        gap_area,
-    );
-
-    let body_y = inner.y.saturating_add(2);
-    let body_h = inner.height.saturating_sub(2);
-    if body_h == 0 {
-        return;
-    }
-    let body_area = Rect {
-        x: inner.x,
-        y: body_y,
-        width: inner.width,
-        height: body_h,
-    };
-
-    let lines = format_grouped_lines(
-        &app.task_panel.snapshot,
-        app.task_panel.scroll,
-        body_h as usize,
-    );
-    let styled: Vec<Line> = lines
-        .into_iter()
-        .map(|l| Line::from(Span::styled(l, row_style)))
-        .collect();
-    frame.render_widget(Paragraph::new(styled), body_area);
+    let ctx = app.render_ctx();
+    agent_tui_kit::render::task_panel::render_task_panel(frame, area, &ctx);
 }
 
 #[cfg(test)]
